Test improvements
Summary:
- make tests in 'ldap' group working with APP_LDAP=0. We're skipping
the backend tests, but we should execute the rest
- code style improvements
- added missing @group annotations
- make Infrastructure tests work with both Cyrus DAV and iRony

Test Plan: `./phpunit tests`

Reviewers: #hkccp_developers, mollekopf

Reviewed By: #hkccp_developers, mollekopf

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D4875

+++ src/tests/BackendsTrait.php
@@ -309,6 +309,13 @@ trait BackendsTrait
     {
         $imap = $this->getImapClient($account);
 
+        // Check the folder existence first, to prevent Cyrus IMAP fatal error when
+        // attempting to delete a non-existing folder
+        $existing = $imap->listMailboxes('', $folder);
+        if (is_array($existing) && in_array($folder, $existing)) {
+            return;
+        }
+
         if (!$imap->deleteFolder($folder)) {
             if (str_contains($imap->error, "Mailbox does not exist")) {
                 // Ignore
